sometimes worker is not available at this stage

+++ standalone/openjscad.js
@@ -833,7 +833,7 @@ OpenJsCad.Processor.prototype = {
       {
       this.worker = OpenJsCad.parseJsCadScriptASync(this.script, paramValues, this.options, function(err, obj) {
         that.processing = false;
-        that.worker.terminate();
+        that.worker && that.worker.terminate();
         that.worker = null;
 
         if(err)
